separated components to different files
