+++ Week2/2048-game/src/GameMessage.css
@@ -1,0 +1,12 @@
+.Game-Message{
+    font-size: large;
+    font-weight: 700;
+}
+
+.Game-Over{
+    color: red;
+}
+
+.Game-Complete{
+    color: green;
+}
